+++ src/main/java/com/roi/cartoh/exception/ResourceNotFoundException.java
@@ -1,7 +1,7 @@
 package com.roi.cartoh.exception;
 
 public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String resource, Long id) {
+    public ResourceNotFoundException(String resource, Integer id) {
         super(resource + " with ID " + id + " not found.");
     }
 }
